login logout and register

+++ base/views.py
@@ -1,10 +1,65 @@
 from django.shortcuts import render, redirect
 from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
+from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib import messages
 from .models import Room, Topic
 from .forms import RoomForm
-from django.db.models import Q
+
 
 # Create your views here.
+
+def loginUser(request):
+    page ='login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, "User does not exist")
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "Username or Password was wrong")
+    context={'page': page}
+    return render(request, "base/login_register.html", context)
+
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+    
+
+def registerUser(request):
+    page ='register'
+    form = UserCreationForm()
+
+    if request.method == 'POST' :
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error for registration')
+
+
+    return render(request, 'base/login_register.html',{'form':form})
 
 
 def room(request, pk):
@@ -12,11 +67,13 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+
 def home(request):
     if request.GET.get('q') != None :
         q = request.GET.get('q')
     
-    else: q = ''
+    else: 
+        q = ''
 
 
     topics = Topic.objects.all()
@@ -30,6 +87,7 @@ def home(request):
     context = {'rooms': rooms, 'topics':topics, 'room_count':room_count}
     return render(request, "base/home.html", context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == "POST":
@@ -41,9 +99,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, "base/room_form.html", context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id = pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host :
+        return HttpResponse('you are not allowed here')
+
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -53,9 +116,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, "base/room_form.html", context)
 
-
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host :
+        return HttpResponse('you are not allowed here')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
